fix(rename python file name): fix rename python file name interface
fix rename python file name interface
