+++ docs/changelog.md
@@ -2204,4 +2204,6 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 0.9.76: add: add stock_us_spot_em, stock_us_hist interface
 
 0.9.77: fix: fix stock_us_hist interface
+
+0.9.78: fix: fix rename python file name interface
 ```
